requirements.txt
distribute==0.6.28
wsgiref==0.1.2
Django==1.4.1
psycopg2==2.4.5
dj-database-url==0.2.0
mysql-python
PIL==1.1.7
django-haystack==1.2.7
pysolr==2.1.0
markdown==2.2.1
python-magic
BeautifulSoup
